Adds "proxies=" as an optional parameter
Youtube frequently blocks IPs orinating a high (or even a moderate) number of requests. This implements the "proxies=" as an optional parameter to scrape Youtube behind a proxy. Proxy is added directly to the created "session" of "requests".

Proxy shall be set in the regular way used by "requests", as in the below example:
```
# proxies = {'https': 'http://101.102.103.104:3128'} 
# or
# proxies = {'https': 'http://username:password@101.102.103.104:3128'}

videos = scrapetube.get_channel('UCzQUP1qoWDoEbmsQxvdjxgQ', proxies=proxies, limit=10, content_type='videos')
```

I AM A NEWBIE IN PYTHON PROGRAMING. MAYBE THERE IS A BETTER WAY TO DO IT OR SOME BUG.
THIS HAS NOT BEEN EXTENSIVELY TESTED.

+++ scrapetube/scrapetube.py
@@ -16,11 +16,12 @@ def get_channel(
     channel_url: str = None,
     channel_username: str = None,
     limit: int = None,
-    sleep: int = 1,
+    sleep: float = 1,
     proxies: dict = None,
     sort_by: Literal["newest", "oldest", "popular"] = "newest",
     content_type: Literal["videos", "shorts", "streams"] = "videos",
 ) -> Generator[dict, None, None]:
+
     """Get videos for a channel.
 
     Parameters:
@@ -187,7 +188,7 @@ def get_video(
 
 
 def get_videos(
-    url: str, api_endpoint: str, selector: str, limit: int, sleep: int, proxies, sort_by: str = None
+    url: str, api_endpoint: str, selector: str, limit: int, sleep: float, proxies, sort_by: str = None
 ) -> Generator[dict, None, None]:
     session = get_session(proxies)
     is_first = True
